# packages/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Bookmark Manager Platform API
  description: |
    Production-ready bookmark management platform with comprehensive organization, archiving, search, and collaboration capabilities.

    ## Features
    - Save and organize web bookmarks with rich metadata
    - Hierarchical collections with custom icons
    - Multi-tag support with advanced filtering
    - Full-text search within saved page content (Pro)
    - Permanent archival with page snapshots (Pro)
    - Highlights and annotations (Pro)
    - Collaboration and sharing (Pro)
    - Cross-device real-time synchronization

    ## Authentication
    The API supports two authentication methods:
    1. **JWT Bearer Tokens** - For direct user authentication
    2. **OAuth 2.0 with PKCE** - For third-party applications

    ## Rate Limiting
    - Free tier: 100 requests per minute per user
    - Pro tier: 500 requests per minute per user
    - Developer API: Rate limits based on application tier

    When rate limits are exceeded, the API returns HTTP 429 with `Retry-After` header.

  version: 1.0.0
  contact:
    name: API Support
    email: support@bookmarkmanager.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.bookmarkmanager.example.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login

    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /oauth/authorize
          tokenUrl: /oauth/token
          scopes:
            read: Read access to bookmarks and collections
            write: Create and update bookmarks and collections
            delete: Delete bookmarks and collections
            admin: Full administrative access
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - timestamp
            - requestId
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: INVALID_REQUEST
            message:
              type: string
              description: Human-readable error message
              example: Invalid request parameters
            details:
              type: object
              description: Additional error context
            timestamp:
              type: string
              format: date-time
              description: ISO 8601 timestamp
            requestId:
              type: string
              description: Unique request identifier for tracing
...
    Bookmark:
      type: object
      required:
        - id
        - owner_id
        - title
        - url
        - type
        - domain
        - created_at
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        collection_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          maxLength: 500
        url:
          type: string
          format: uri
        excerpt:
          type: string
          nullable: true
        content_snapshot_path:
          type: string
          nullable: true
          description: Path to archived snapshot (Pro)
        content_indexed:
          type: boolean
          default: false
        type:
          type: string
          enum: [article, video, image, file, document]
        domain:
          type: string
          maxLength: 255
        cover_url:
          type: string
          format: uri
          nullable: true
        is_duplicate:
          type: boolean
          default: false
        is_broken:
          type: boolean
          default: false
        custom_order:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        highlights:
          type: array
          items:
            $ref: '#/components/schemas/Highlight'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Tag:
      type: object
      required:
        - id
        - owner_id
        - name
        - normalized_name
      properties:
        id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 100
        normalized_name:
          type: string
          maxLength: 100
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: '#FF5733'
        created_at:
          type: string
          format: date-time
    Highlight:
      type: object
      required:
        - id
        - bookmark_id
        - owner_id
        - text_selected
        - color
      properties:
        id:
          type: string
          format: uuid
        bookmark_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        text_selected:
          type: string
          description: Selected text content
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          default: '#FFFF00'
        annotation_md:
          type: string
          nullable: true
          description: Markdown annotation
        position_context:
          type: object
          description: Position information for re-highlighting
        snapshot_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /bookmarks:
    get:
      tags:
        - Bookmarks
      summary: List bookmarks with filtering
      operationId: listBookmarks
      security:
        - bearerAuth: []
        - oauth2: [read]
      parameters:
        - name: collection_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by collection
        - name: tags
          in: query
          schema:
            type: array
            items:
              type: string
          description: Filter by tags (comma-separated)
        - name: type
          in: query
          schema:
            type: string
            enum: [article, video, image, file, document]
        - name: domain
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of bookmarks
          content:
            application/json:
              schema:
                type: object
                properties:
                  bookmarks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bookmark'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
              description: Seconds until rate limit resets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Bookmarks
      summary: Create a new bookmark
      operationId: createBookmark
      security:
        - bearerAuth: []
        - oauth2: [write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                title:
                  type: string
                  maxLength: 500
                excerpt:
                  type: string
                collection_id:
                  type: string
                  format: uuid
                tags:
                  type: array
                  items:
                    type: string
                note:
                  type: string
            example:
              url: https://example.com/article
              title: Interesting Article
              excerpt: This is a great article about...
              collection_id: 550e8400-e29b-41d4-a716-446655440000
              tags:
                - javascript
                - web-development
      responses:
        '201':
          description: Bookmark created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [processing, completed]
                  bookmark:
                    $ref: '#/components/schemas/Bookmark'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...